+++ src/ct/v1.rs
@@ -1,4 +1,8 @@
+use deku::prelude::*;
 use serde::{Deserialize, Serialize};
+use x509_parser::prelude::*;
+
+use crate::CTLogError;
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct AddChainResponse {
@@ -96,4 +100,107 @@ pub struct GetEntryAndProofResponse {
     /// An array of base64-encoded Merkle Tree nodes proving
     /// the inclusion of the chosen certificate.
     pub audit_path: Vec<String>,
+}
+
+#[derive(Debug, PartialEq, Eq, Clone, DekuRead)]
+#[deku(type = "u16", endian = "big")]
+pub enum LogEntryType {
+    X509Entry = 0,
+    PrecertEntry = 1,
+}
+
+#[derive(Debug, Clone, DekuRead)]
+pub struct ASN1Cert<'a> {
+    #[deku(bytes = 3, endian = "big")]
+    pub length: u32,
+    #[deku(
+        count = "length",
+        map = "|v: &'a [u8]| -> Result<_, DekuError> { Ok(X509Certificate::from_der(v).unwrap().1) }"
+    )]
+    pub certificate: X509Certificate<'a>,
+}
+
+#[derive(Debug, Clone, DekuRead)]
+pub struct ASN1CertChain<'a> {
+    #[deku(bytes = 3, endian = "big")]
+    pub length: u32,
+    #[deku(bytes_read = "length")]
+    pub certificates: Vec<ASN1Cert<'a>>,
+}
+
+#[derive(Debug, Clone, DekuRead)]
+pub struct PrecertChainEntry<'a> {
+    pub pre_certificate: ASN1Cert<'a>,
+    pub precertificate_chain: ASN1CertChain<'a>,
+}
+
+#[derive(Debug, PartialEq, Eq, Clone, DekuRead)]
+#[deku(type = "u8")]
+pub enum Version {
+    V1 = 0,
+}
+
+#[derive(Debug, Clone, DekuRead)]
+pub struct PreCert<'a> {
+    pub issuer_key_hash: [u8; 32],
+    #[deku(bytes = 3, endian = "big")]
+    pub length: u32,
+    #[deku(
+        count = "length",
+        map = "|v: &'a [u8]| -> Result<_, DekuError> { println!(\"{:?}\", v);Ok(TbsCertificate::from_der(v).unwrap().1) }"
+    )]
+    pub tbs_certificate: TbsCertificate<'a>,
+}
+
+#[derive(Debug, Clone, DekuRead)]
+pub struct CtExtensions {
+    #[deku(endian = "big")]
+    pub length: u16,
+    #[deku(count = "length")]
+    pub extensions: Vec<u8>,
+}
+
+#[derive(Debug, PartialEq, Eq, Clone, DekuRead)]
+#[deku(type = "u8")]
+pub enum MerkleLeafType {
+    TimestampedEntry = 0,
+}
+
+#[derive(Debug, Clone, DekuRead)]
+#[deku(ctx = "entry_type: LogEntryType", id = "entry_type")]
+pub enum TimestampedEntrySignedInner<'a> {
+    #[deku(id = "LogEntryType::X509Entry")]
+    X509(ASN1Cert<'a>),
+    #[deku(id = "LogEntryType::PrecertEntry")]
+    Precert(PreCert<'a>),
+}
+
+#[derive(Debug, Clone, DekuRead)]
+pub struct TimestampedEntry<'a> {
+    #[deku(endian = "big")]
+    pub timestamp: u64,
+    pub entry_type: LogEntryType,
+    #[deku(ctx = "entry_type.clone()")]
+    pub signed_entry: TimestampedEntrySignedInner<'a>,
+    pub extensions: CtExtensions,
+}
+
+#[derive(Debug, Clone, DekuRead)]
+pub struct MerkleTreeLeaf<'a> {
+    pub version: Version,
+    pub leaf_type: MerkleLeafType,
+    pub timestamped_entry: TimestampedEntry<'a>,
+}
+
+#[derive(Debug, Clone)]
+pub enum DecodedEntryInner<'a> {
+    X509(ASN1CertChain<'a>),
+    Precert(PrecertChainEntry<'a>),
+}
+
+/// A structure representing a log entry (parsed from the response of /ct/v1/get-entries).
+#[derive(Debug, Clone)]
+pub struct DecodedEntry<'a> {
+    pub leaf: MerkleTreeLeaf<'a>,
+    pub extra_data: DecodedEntryInner<'a>,
 }